根据计算，修改存储类型便于数据库存储（效率，空间和便捷性）

+++ src/main/java/ai/techfin/tradesystem/domain/Placement.java
@@ -3,11 +3,8 @@ package ai.techfin.tradesystem.domain;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import javax.persistence.Embedded;
+import javax.persistence.*;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.Instant;
 
 @Embeddable
@@ -19,7 +16,7 @@ public class Placement {
     private Stock stock;
 
     @Column(nullable = false)
-    private BigInteger quantity;
+    private Long quantity;
 
     @Column(nullable = false, precision = 21, scale = 2)
     private BigDecimal price;
@@ -27,7 +24,7 @@ public class Placement {
     @Column(name = "created_at", nullable = false)
     private Instant createdAt = Instant.now();
 
-    public Placement(Stock stock, BigInteger quantity, BigDecimal price) {
+    public Placement(Stock stock, Long quantity, BigDecimal price) {
         this.stock = stock;
         this.quantity = quantity;
         this.price = price;
@@ -39,6 +36,6 @@ public class Placement {
         return price;
     }
 
-    public BigInteger getQuantity() { return quantity; }
+    public Long getQuantity() { return quantity; }
 
 }
